rename view files
