+++ src/main.js
@@ -1,12 +1,12 @@
-import {createHeaderProfileTemplate} from "./view/headerProfile";
+import {createHeaderProfileTemplate} from "./view/header-profile";
 import {createNavigationTemplate} from "./view/navigation";
 import {createSortTemplate} from "./view/sort";
-import {createFilmsMainListTemplatae} from "./view/filmsMainList";
-import {createFilmsListContainerTemplate} from "./view/filmsListContainer";
-import {createFilmCardTemplatae} from "./view/filmCard";
-import {createShowMoreButtonTemplate} from "./view/showMoreButton";
-import {createTopRatedListTemplatae} from "./view/topRatedList";
-import {createMostCommentedListTemplatae} from "./view/mostCommentedList";
+import {createFilmsMainListTemplatae} from "./view/films-main-list";
+import {createFilmsListContainerTemplate} from "./view/films-list-container";
+import {createFilmCardTemplatae} from "./view/film-card";
+import {createShowMoreButtonTemplate} from "./view/show-more-button";
+import {createTopRatedListTemplatae} from "./view/top-rated-list";
+import {createMostCommentedListTemplatae} from "./view/most-commented-list";
 import {createStatisticsTemplate} from "./view/statistics";
 
 const FILMS_COUNT = 5;
