+++ src/view/films-list-container.js
