+++ src/view/films-main-list.js
