+++ src/view/show-more-button.js
