+++ src/view/top-rated-list.js
